fix: fold submitWorkResponseFromResult to satisfy pkg-file-count gate
Merge submit_work_response.go into submit_work_items.go so pkg/api stays
within the 15-file limit after story 001 added the response mapper.
